Move actions to app/appActions
