+++ src/app/AppSettings.js
@@ -4,7 +4,7 @@ import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import { FormattedMessage } from 'react-intl';
 import { getLocale } from '../reducers';
-import * as actions from '../actions';
+import * as actions from './appActions';
 
 @connect(
   state => ({
